Вынесет формы входа и регистрации в компоненты
